feat: floating CTA, nav improvements, dead link fixes, case studies SEO
- Add FloatingCTA component: sticky contact/connect button that appears
  after scrolling past the hero and hides near the #connectwithme section.
  Includes email shortcut and smooth-scroll behavior.

- Fix Navigation logo: wrap in proper Next.js Link for correct semantics,
  accessibility, and SEO (was a non-navigable div with onClick only).

- Add Case Studies link to navigation with active state highlighting.
  Also import usePathname to support active states on sub-pages (/case-studies,
  /experience).

- Fix all dead href="#" links:
  - Testimonials: 12 LinkedIn buttons now point to Brandon's LinkedIn
    recommendations page instead of href="#".
  - IndustryCollaboration: Resume and Executive Brief buttons now use
    direct PDF hrefs with target="_blank" instead of JavaScript-driven
    href="#" downloads.

- Convert case-studies page to server component + SEO metadata:
  - Extract client rendering into CaseStudiesContent.tsx ('use client')
  - page.tsx becomes a server component exporting full Metadata (title,
    description, keywords, OG, Twitter, canonical) for Google indexing.
  - Also update back-link to go to "/" (Home) instead of a section anchor.
  - Add id attributes to each case study article for anchor deep-linking.
